tbvision/requirements.txt
# === Web Framework ===
fastapi==0.129.1            
uvicorn[standard]      
starlette==0.52.1
pydantic==2.12.5
pydantic-settings==2.13.1
langchain-community==0.4.1

# === Vector Database ===
qdrant-client==1.17.0

# === Embeddings / Models ===
sentence-transformers==5.2.3

# === LLM / Providers ===
google-genai==1.64.0
transformers==5.2.0
mistralai==1.12.4

# === Utilities ===
python-multipart==0.0.22
pypdf

numpy
opencv-python
python-dotenv
